# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperrors.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  apperrors.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  model.Driver:
    properties:
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
      status:
        $ref: '#/definitions/model.DriverStatus'
    type: object
  model.DriverInput:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
      status:
        type: integer
    type: object
  model.DriverStatus:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.TokenResponse:
    properties:
      access_token:
        type: string
    type: object
...
  model.UserCreate:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  model.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: luchitstimofei@yandex.ru
    name: Luchits Timofei
    url: http://example.com
  description: This is a server for order private bus.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ordering Bus API
  version: "1.0"
paths:
...
  /admin/driver:
    post:
      consumes:
      - application/json
      description: Создает новую запись о водителе (требуются права администратора)
      parameters:
      - description: Данные водителя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.DriverInput'
      produces:
      - application/json
      responses:
        "201":
          description: Водитель успешно создан
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Создать нового водителя (только админ)
      tags:
      - admin/driver
  /admin/driver/{id}:
    delete:
      description: Удаляет запись о водителе по ID (требуются права администратора)
      parameters:
      - description: ID водителя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Водитель успешно удален
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Удалить водителя (только админ)
      tags:
      - admin/driver
    put:
      consumes:
      - application/json
      description: Обновляет информацию о водителе по ID (требуются права администратора)
      parameters:
      - description: ID водителя
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные водителя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.DriverInput'
      produces:
      - application/json
      responses:
        "200":
          description: Данные водителя обновлены
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Обновить данные водителя (только админ)
      tags:
      - admin/driver
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Вход пользователя в систему
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/model.TokenResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/model.UserResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
...
  /driver:
    get:
      description: Возвращает список всех водителей в системе
      produces:
      - application/json
      responses:
        "200":
          description: Список водителей
          schema:
            items:
              $ref: '#/definitions/model.Driver'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить список всех водителей
      tags:
      - drivers
  /driver/{id}:
    get:
      description: Возвращает информацию о водителе по его идентификатору
      parameters:
      - description: ID водителя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные водителя
          schema:
            $ref: '#/definitions/model.Driver'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить водителя по ID
      tags:
      - drivers
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"